+++ core/src/main/java/de/pxav/kelp/core/event/listener/EventHandlerRegistration.java
@@ -1,4 +1,4 @@
-package de.pxav.kelp.core.event;
+package de.pxav.kelp.core.event.listener;
 
 import com.google.common.base.Preconditions;
 import com.google.inject.Inject;
